Database Hotfix1

+++ database/migrations/2019_11_16_095704_create_schedules_table.php
@@ -17,7 +17,7 @@ class CreateSchedulesTable extends Migration
             $table->bigIncrements('id');
             $table->dateTime('date_time_start');
             $table->integer('duration');
-            $table->integer('room_id');
+            $table->unsignedBigInteger('room_id');
             $table->foreign('room_id')->references('id')->on('rooms')->onDelete('cascade');
             $table->timestamps();
         });
